fan_out/pyproject.toml
[project]
name = "fan_out"
version = "0.1.0"
description = "Fan-out request flow Async RPC example app."
authors = [
    { name = "flossypurse", email = "flossypurse@gmail.com" }
]
dependencies = [
    "flask>=3.1.0",
    "resonate-sdk==0.4.12",
]

readme = "README.md"
requires-python = ">=3.12"

[project.scripts]
foo = "src.service_foo:main"
bar = "src.service_bar:main"
baz = "src.service_baz:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
